Break client and series classes into their own files
